incusd/network: Remove duplicated logic
